Process CSV files via background jobs
This commit setups Active Job and Active Storage in order to process a
POSTed CSV file to the addresses API endpoint.

When the endpoint is hit, we first create a new CsvFile model, attaches
the file, and then enqueues a job to process the CSV file.

The BaseCsvImporter base service model to accept an Active Storage
attachment instead of a filename, and uses that to open and process the
CSV file.

The configuration of Active Job and Active Storage is the base
configuration outlined in the official Rails guides. In development, we
are storing CSVs to disk for Active Storage, and we are using Solid Queue as
the backend for Active Job.